+++ app/[usernameSlug]/profile-tabs.tsx
@@ -10,10 +10,13 @@ import { Investments } from './user-investments'
 import { Projects } from './user-projects'
 import { RichContent } from '@/components/editor'
 import { BalanceDisplay } from './balance-display'
-import { calculateUserBalance, calculateUserSpendableFunds } from '@/utils/math'
+import {
+  calculateUserBalance,
+  calculateUserSpendableFunds,
+  categorizeTxn,
+} from '@/utils/math'
 import { Txn, FullTxn } from '@/db/txn'
-import { sortBy } from 'lodash'
-import { BANK_ID } from '@/db/env'
+import { calculateWithdrawBalance } from '@/utils/math'
 import { DonateBox } from '@/components/donate-box'
 import { OutgoingDonationsHistory } from './user-donations'
 
@@ -149,117 +152,6 @@ export function ProfileTabs(props: {
   } else {
     return null
   }
-}
-
-function calculateWithdrawBalance(
-  txns: Txn[],
-  bids: Bid[],
-  userId: string,
-  balance: number,
-  accreditationStatus: boolean
-) {
-  let nonWithdrawBalance = 0
-  const sortedTxns = sortBy(txns, 'created_at')
-  sortedTxns.forEach((txn) => {
-    const txnType = categorizeTxn(txn, userId)
-    if (txnType === 'non-dollar') return
-    if (txn.to_id === userId) {
-      if (!txnWithdrawable(txnType, accreditationStatus)) {
-        nonWithdrawBalance += txn.amount
-      }
-    } else {
-      if (txnCharitable(txnType, accreditationStatus)) {
-        nonWithdrawBalance -= txn.amount
-      }
-    }
-    nonWithdrawBalance = Math.max(nonWithdrawBalance, 0)
-  })
-  bids.forEach((bid) => {
-    if (bid.status === 'pending' && bid.type === 'buy') {
-      nonWithdrawBalance += bid.amount
-    }
-  })
-  return Math.max(balance - nonWithdrawBalance, 0)
-}
-
-type TxnType =
-  | 'incoming cash transfer'
-  | 'outgoing cash transfer'
-  | 'share purchase'
-  | 'share sale'
-  | 'own-project share sale'
-  | 'withdraw'
-  | 'deposit'
-  | 'incoming project donation'
-  | 'outgoing project donation'
-  | 'non-dollar'
-
-function categorizeTxn(txn: FullTxn, userId: string) {
-  if (txn.token === 'USD') {
-    if (txn.to_id === userId) {
-      if (txn.project) {
-        if (txn.projects?.creator === userId) {
-          if (txn.bundle) {
-            return 'own-project share sale'
-          } else {
-            return 'incoming project donation'
-          }
-        } else {
-          return 'share sale'
-        }
-      } else {
-        if (txn.from_id === BANK_ID) {
-          return 'deposit'
-        } else {
-          return 'incoming cash transfer'
-        }
-      }
-    } else {
-      if (txn.project) {
-        if (txn.bundle) {
-          return 'share purchase'
-        } else {
-          return 'outgoing project donation'
-        }
-      } else {
-        if (txn.to_id === BANK_ID) {
-          return 'withdraw'
-        } else {
-          return 'outgoing cash transfer'
-        }
-      }
-    }
-  } else {
-    return 'non-dollar'
-  }
-}
-
-// For incoming txns only
-function txnWithdrawable(txnType: TxnType, accreditationStatus: boolean) {
-  if (txnType === 'incoming cash transfer') {
-    return false
-  }
-  if (
-    txnType === 'share sale' ||
-    (txnType === 'deposit' && !accreditationStatus)
-  ) {
-    return false
-  }
-  return true
-}
-
-// For outgoing txns only
-function txnCharitable(txnType: TxnType, accreditationStatus: boolean) {
-  if (
-    txnType === 'outgoing cash transfer' ||
-    txnType === 'outgoing project donation'
-  ) {
-    return true
-  }
-  if (txnType === 'share purchase' && !accreditationStatus) {
-    return true
-  }
-  return false
 }
 
 export type Investment = {
