+++ app/withdraw/page.tsx
@@ -1,8 +1,10 @@
 import { createServerClient } from '@/db/supabase-server'
 import { WithdrawalSteps } from './withdrawal-steps'
-import { getUser } from '@/db/profile'
+import { getProfileAndBidsById, getUser } from '@/db/profile'
 import { STRIPE_SECRET_KEY } from '@/db/env'
+import { calculateUserBalance, calculateWithdrawBalance } from '@/utils/math'
 import Stripe from 'stripe'
+import { getFullTxnsByUser } from '@/db/txn'
 
 const stripe = new Stripe(STRIPE_SECRET_KEY as string, {
   apiVersion: '2022-11-15',
@@ -20,6 +22,16 @@ export default async function WithdrawPage() {
       </div>
     )
   }
+  const profile = await getProfileAndBidsById(supabase, user.id)
+  const txns = await getFullTxnsByUser(supabase, user.id)
+  const balance = calculateUserBalance(txns, user.id)
+  const withdrawBalance = calculateWithdrawBalance(
+    txns,
+    profile.bids,
+    user.id,
+    balance,
+    profile.accreditation_status
+  )
   const { data } = await supabase
     .from('profiles')
     .select('stripe_connect_id')
@@ -37,6 +49,7 @@ export default async function WithdrawPage() {
       <WithdrawalSteps
         account={account}
         userId={user.id}
+        withdrawBalance={withdrawBalance}
         loginLink={loginLink}
       />
     </div>
